requirements.txt
pandas>=2.2
numpy>=1.26
plotly>=5.22
streamlit>=1.36
python-dotenv>=1.0
duckdb>=1.0
statsmodels>=0.14
scikit-learn>=1.4
openai>=1.40
matplotlib>=3.8
pyarrow>=16.0
